ROC-2046: refactoring, renaming files
