+++ src/main/features/ccj/routes/their-details.ts
@@ -7,7 +7,7 @@ import { PartyDetails } from 'app/forms/models/partyDetails'
 import { FormValidator } from 'app/forms/validation/formValidator'
 import { Form } from 'app/forms/form'
 import User from 'app/idam/user'
-import { DraftCCJService } from 'ccj/draft/DraftCCJService'
+import { DraftCCJService } from 'ccj/draft/draftCCJService'
 import { Address } from 'forms/models/address'
 import { PartyDetailsFactory } from 'forms/models/partyDetailsFactory'
 import { TheirDetails } from 'claims/models/details/theirs/theirDetails'
