+++ src/main/features/testing-support/index.ts
@@ -4,7 +4,7 @@ import * as path from 'path'
 
 import { AuthorizationMiddleware } from 'idam/authorizationMiddleware'
 import { RouterFinder } from 'common/router/routerFinder'
-import { buildURL } from 'utils/CallbackBuilder'
+import { buildURL } from 'utils/callbackBuilder'
 import { Paths } from 'app/paths'
 
 function defendantResponseRequestHandler (): express.RequestHandler {
