+++ src/test/features/ccj/guards/ccjGuard.ts
@@ -7,7 +7,7 @@ import { mockReq as req, mockRes as res } from 'sinon-express-mock'
 
 import { Paths } from 'dashboard/paths'
 
-import { DraftCCJ } from 'ccj/draft/DraftCCJ'
+import { DraftCCJ } from 'ccj/draft/draftCCJ'
 import { Defendant } from 'drafts/models/defendant'
 import { PartyDetails } from 'forms/models/partyDetails'
 import { CCJGuard } from 'ccj/guards/ccjGuard'
